LCGDM-1376: ensure a specific starting order of xrd daemons
the data structure for the xrootd instance options
had to be changed from a hash with all options
to a list of hashes.

The list is ordered and guarantees the startup of
the daemons in the right order, while the hashes
hold multiple instances of the same type.

Only works with a change in module dmlite.